requirements.txt
requests
python-dotenv
apify-client
huggingface_hub python-dotenv
pydantic